Facebook login/register v2.4+ compatibility
Changed the way the params are passed to the API . This is backwards compatible with version v2.3 and older. Also apps created today will only be available for today's API version and up, so in this case v2.5. So the current code won't work for newer facebook apps (like mine)

https://developers.facebook.com/docs/apps/changelog/#v2_4
https://developers.facebook.com/docs/graph-api/using-graph-api/v2.5